feat: onboarding (#30)
When no data has been ingested (in reality, when loading app names is
empty), we display a dialog with some onboarding info.


![image](https://github.com/grafana/pyroscope-app-plugin/assets/6951209/282e085a-9b31-47d8-a324-d5da0a884790)


* I ended up using `grafana/ui` Dialog instead of pyroscope's. Reasons
are:
* Modal is something that is rendered outside the app plugin (aka
grafana), so it meshes better with grafana's theming
* We get a feeling on how it will be when migrating all of our
components to @grafana/ui (which we want to do at some point in the
future)

Right now it's linking to
https://grafana.com/docs/phlare/latest/configure-client/, ideally would
be nice to link to that person's pyroscope stack info. Eg
https://grafana.com/orgs/pyroscope/hosted-profiles/213487, but I don't
know how to figure out that url yet.

Also differently from Pyroscope Cloud, here we ALWAYS show the modal if
there's no data. IE there's no "Dismiss forever" button. This is
aggressive to get people to start ingesting data!

It doesn't keep listening for ingested data for the dialog to be closed
automatically. This is to be on the safe side and not overload the
backend.

TODO:
- [x] tests
- [x] check if light mode is working

---------

Co-authored-by: Dmitry Filimonov <dmitry.filimonov@grafana.com>