fix: handle inconsistent date formats (hack) (#60)
I had fixed an issue where the `Sync Timelines` button was resulting in
dates that were off by a factor of 1000, and that fix broke the
pyroscope navigational drag selector forcing the dates to the 1970s
(also off by a factor of 1000).

It appears that pyroscope is providing `from` and `until` in both
seconds and milliseconds depending on which action initiates the change.
Pyroscope on its own seems able to handle this duality, but the
translation code in the app plugin was doing it wrong one way or
another.

This PR provides a hack to work around the discrepancy. If the date
resolves to >= year 2000, we assume the date format is already in
milliseconds and requires no translation. Otherwise, we assume it is in
seconds and multiply by 1000. I will investigate the root cause on the
pyroscope code and try to find out why the app plugin was receiving two
separate formats, but this workaround hack should resolve this issue for
now.