refactor: toolbar state management via context to reduce jitter (#66)
To support Grafanafied components such as the recently added Toolbar
replacement, we need to maintain a Grafana version of the app/profile
selection state, and the selected time range.

Previously, on render, the Toolbar would have to derive this Grafana
translation of the state from the Pyroscope app state, which would take
cause a visible jitter in the toolbar as things get resolved. This
jitter would happen each time the page changed, since the Pyroscope
pages contain the Toolbar as part of their rendering.

This PR moves the Grafana version of the state into an outer context
wrapper to maintain it between page changes. The Toolbar now draws from
this context, keeping it up to date. This refactor also centralizes
state control and translation between Grafana and Pyroscope in a more
neat fashion (subjectively).