fix: remove plugin link extension declaration to avoid historical clashes with Grafana (#447)
Due to historical issues with this plugin being preloaded in October
2023 causing clashes with other aspects of Grafana, we disabled the
preload. The clashes were related to "og" pyroscope app code making
global changes:
- Competing URL syncing strategies, which invalidated Grafana dashboard
URL parameters
- Overriding common panel rendering code, which broke a few specific
Grafana plugins

The clashes would occur when the plugin's code was loaded (since it also
loaded pyroscope "og" code as well). When preload was set to true,
pyroscope code would load immediately, causing any of these clash bugs
to always occur. When we set preload to false, pyroscope code would only
load when the pyroscope plugin was accessed -- thus issue could be
worked around. The identified clashing issues were resolved, but for
safety, we decided to disabled the "preload" capability altogether in
case there were any undiscovered clash-type bugs, and thus the extension
capability of this plugin has been deactivated (since it required
preload to be true in order for them to register).

The following recent PR activates the extension again through a new
mechanism.
- https://github.com/grafana/explore-profiles/pull/437

If I understand correctly, this will cause the plugin's code to load
whenever the extension point requests extensions (i.e., when the user
accesses the Explore page on a pyroscope datasource). If there are any
existing clash-type bugs, there is now this second way of activating
them.

Note: As we migrate away from directly using "OG" pyroscope code, the
risk of undiscovered clash-type bugs is reduced.

Note: If there are still undiscovered clash-type bugs, they can still be
avoided by not accessing the pyroscope plugin or datasource query
editor.

Here are a list of issues and PRs related to the extension, and some of
the clash-type bugs that were discovered:
- https://github.com/grafana/explore-profiles/issues/171
- https://github.com/grafana/explore-profiles/issues/169
- https://github.com/grafana/explore-profiles/issues/168
- https://github.com/grafana/explore-profiles/pull/163
- https://github.com/grafana/explore-profiles/pull/161

Here are some options to consider:

A) Keep the effects of the
[PR](https://github.com/grafana/explore-profiles/pull/437), which
restore this link extension from the pyroscope datasoruce to the
profiles app.
- This requires the URL that is generated from the data source to to
still be valid (i.e., we didn't change the structure of the app)
- This adds a potential loading point where clash-type bugs could be
activated; when the user accesses the pyroscope datasource
- To choose this option, simply close this PR.

B) Restore the previous behavior of deactivating the extension
- We can consider bringing back the "extensions" declaration that this
PR removes after we properly evaluate that everything is safe and
functional
- Simply merge this PR, and make backlog issue to reverse its effect
when we are ready for it

C) Completely remove/rethink the extension point that was added to the
pyroscope datasource.
- We wanted to use that link to bring further awareness to the profiles
app. Is this still a priority? Maybe we remove the extension point
altogether?
- This requires merging this PR, deleting the extension code from this
repo, and the grafana repo.

D) Consider a general purpose explore link to explore app that is not
specifically tied to pyroscope.