Simplify showing applied filters
Showing a "chiclet" suggests it could be an editable (especially with blue call to action). To reduce confusion we can simply show filters as a string. This should simplify UI and code.

The labels and descriptions are also a bit confusing and can be improved. I changed them slightly but it still needs revising.