feat(Metrics): Basic UI for experimental metrics from profiles (#462)
* Sketch out exported metrics list view

* yarn lint:fix

* Add export metrics option to flame graph context menu

* Create modal and integrate flame graph menu

* Minor naming changes

* Load label names from the backend server

* Update view metrics page

* Use Table component to layout metric list

* Use "Create metric" instead of "Export metric"

* Use 'Create' instead of 'Save'

* Show success banner when save is successful

* Save to tenant settings backend

* Include filters when creating metric

* Load metrics from tenant settings

* Implement deleting metrics

* Make it clear the data source is fake

* Autofocus the name field

* Use the profile id when saving profile type

* Improve table layout

* Give rows more height

* Add form validation

* Use better defaults for label select

Co-authored-by: Marc M. <146180665+grafakus@users.noreply.github.com>

* Autofocus the name field

* Use the profile id when saving profile type

* Improve table layout

* Give rows more height

* Add form validation

* Use better defaults for label select

Co-authored-by: Marc M. <146180665+grafakus@users.noreply.github.com>

* Rename metrics route constant

* Remove metrics from side bar

* Improve error feedback with metric name field

* Create readonly chiclets and revamp the form layout

* Reorder imports

* Rename references of "metric" to "recording rule"

* Simplify showing applied filters

Showing a "chiclet" suggests it could be an editable (especially with blue call to action). To reduce confusion we can simply show filters as a string. This should simplify UI and code.

The labels and descriptions are also a bit confusing and can be improved. I changed them slightly but it still needs revising.

* Simplify the code around opening the modal

Creating additional hook feels ab bit redundant

* Reset the modal after cancel

Otherwise, cancelling the modal and opening again keeps the previous state for the name of the metric

* Allow sending no labels

Without this change it fails silently making UI look like it's not responding

* Inline error icon

If the message is a div it's pushed below the alert icon taking more space

* Shorten import

Minor formatting

* Validate onBlur

This will give faster feedback

* Add a setting to enable metrics from profiles

This way we can merge something that may not be ready yet for development purposes. We can remove the setting once the feature is stable

* Add a back button to all sub-views

This way it's easier for the user to exit the view

* Use InteractiveTable

This way the table is more responsive, doesn't need data frame as data (Table component is more useful for rendering data frames)

* Refetch created rules after a delete

This is to reflect deletion

* Reset the form when the modal is closed

This way we always get a clean state on a new modal

* Validate form on change

This way there's no validation error shown when user clicks Cancel straight after opening the modal

* Add sorting and pagination to the table

* Hide the "create" button when feature is disabled

* Disable lazy loading

This needs to be added in https://github.com/grafana/explore-profiles/issues/423

* Remove redundant key

* Fix a unit test

* Fix tests

* Create an e2e test for settings

* Use RecordingRules api

* Pass filters when creating a rule

* Display error message

* Reuse back button

* Fix a typo

* Move RecordingRuleViewModel

* Remove redundant any

* Cache buster for dev testing

* (revert) Cache buster for dev testing

* Add feature toggle check

* Create a trash icon with modal to delete recording rules

* Skip tests since the feature toggle is missing for now

---------

Co-authored-by: bryanhuhta <bryan.huhta@grafana.com>
Co-authored-by: Bryan Huhta <32787160+bryanhuhta@users.noreply.github.com>
Co-authored-by: Marc M. <146180665+grafakus@users.noreply.github.com>